// src/types/CardFanficListagemType.ts
export interface CardFanficListagemProps {
  pathCapa: string;
  altCapa?: string;
  titulo: string;
  pathClassificacao: string;
  altClassificação: string;
  autore: string;
  sinopse: string;
  comentario: number;
  capitulo: string;
  data: string;
  linkFanfic: string;
  linkAutore: string;
  }